hyper_chat/requirements.txt
aioredis==1.1.0
amqp==2.3.2
asgi-redis==1.4.3
asgiref==2.3.2
async-timeout==2.0.1
asyncio==3.4.3
atomicwrites==1.1.5
attrs==18.1.0
autobahn==18.6.1
Automat==0.7.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.4.16
channels==2.1.2
channels-redis==2.2.1
chardet==3.0.4
constantly==15.1.0
daphne==2.2.1
dj-database-url==0.5.0
Django==2.0.7
django-crontab==0.7.1
django-fcm==0.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
kombu==4.2.1
more-itertools==4.2.0
msgpack==0.5.6
msgpack-python==0.5.6
pipenv==2018.7.1
pluggy==0.6.0
psycopg2==2.7.5
py==1.5.4
pyfcm==1.4.5
PyHamcrest==1.9.0
PyJWT==1.6.4
pytest==3.6.3
pytest-asyncio==0.8.0
pytest-django==3.3.2
pytz==2018.5
redis==2.10.6
requests==2.19.1
requests-toolbelt==0.8.0
six==1.11.0
Twisted==18.7.0
txaio==2.10.0
urllib3==1.23
vine==1.1.4
virtualenv==16.0.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
zope.interface==4.5.0
